Add tests directory and move tokens into their own file
Lots of warnings from unused functions need fixing.
Minimal parts of the library are exposed.
Tests use only exposed functions/types.
