+++ src/lib.rs
@@ -1,2 +1,7 @@
-pub mod lexer;
-pub mod parser;
+mod lexer;
+mod parser;
+mod token;
+
+pub use lexer::Lexer;
+pub use parser::Parser;
+pub use token::TokenKind;
